Cinder: improve unit tests factorization
There is no need to pass around the "is_multi_store" parameter when it
can be an attribute of the test class, thus allowing us to have more
code in the base class.
